+++ scriptabit/habitica_service.py
@@ -85,17 +85,43 @@ class HabiticaService(object):
         response.raise_for_status()
         return response.json()['data']
 
-    def create_task(self, task):
-        """Create a new task.
+    def upsert_task(self, task):
+        """Upserts a task.
+
+        Existing tasks will be updated, otherwise a new task will be created.
+        If the task has an alias it will be used for the search, otherwise the
+        id will be used. If neither an id or an alias is provided, then an
+        error is raised.
 
         Args:
-            task (dict): The new task.
+            task (dict): The task.
 
         Returns: dict: The new task as returned from the server.
+
+        Raises:
+            InvalidHabiticaDataError
         """
-        response = self.__post('tasks/user', task)
-        response.raise_for_status()
-        return response.json()['data']
+        # do we use the task ID or alias?
+        key = task['alias'] if 'alias' in task else task.get('_id', None)
+        if not key:
+            raise InvalidHabiticaDataError(
+                'The task must specify an id or alias')
+
+        # Does the task already exist?
+        response = self.__get('tasks/{key}'.format(key=key))
+        if response.status_code == requests.codes.ok:
+            # update
+            logging.getLogger(__name__).debug('task %s exists, updating', key)
+            response = self.__put('tasks/{0}'.format(key), task)
+            response.raise_for_status()
+            return response.json()['data']
+        else:
+            # create
+            logging.getLogger(__name__).debug(
+                'task %s not found, creating', key)
+            response = self.__post('tasks/user', task)
+            response.raise_for_status()
+            return response.json()['data']
 
     # TODO: I don't think the API lets me set partial user objects in this way.
     # So I could get the entire user structure, swap the stats for the argument